separate the http and non http probe tests
